+++ core/src/com/levicore/test/TestState.java
@@ -23,6 +23,8 @@ public class TestState extends State {
                 System.out.print("haha");
             }
         }));
+
+
     }
 
     @Override