+++ core/src/org/xguzm/pathfinding/PathFinderOptions.java
@@ -1,0 +1,6 @@
+package org.xguzm.pathfinding;
+
+public abstract class PathFinderOptions {
+	/** The {@link org.xguzm.pathfinding.Heuristic} to calculate the distance from one node to another */
+	public Heuristic heuristic;
+}