+++ desktop/src/com/levicore/libgdxcore/desktop/DesktopLauncher.java
@@ -3,10 +3,11 @@ package com.levicore.libgdxcore.desktop;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
 import com.levicore.libgdxcore.LibgdxCore;
+import com.levicore.test.MapLoadingTest;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		new LwjglApplication(new LibgdxCore(), config);
+		new LwjglApplication(new MapLoadingTest(), config);
 	}
 }